fix: test cleanup
