+++ test/unit/src/utils/multisigIntegration.test.ts
@@ -1,4 +1,4 @@
-import { isKeyChange, cleanMultiSigPermissions } from '../../../../src/utils/multisig'
+import { isKeyChange, cleanMultiSigPermissions, isNonKeyChange } from '../../../../src/utils/multisig'
 import { expect, describe, test } from '@jest/globals'
 import { ChangeConfig } from '../../../../src/shardeum/shardeumTypes'
 import { DevSecurityLevel } from '@shardeum-foundation/core'
@@ -10,19 +10,24 @@ const mockCycle = {
   // Add other required properties as needed
 };
 
-const setup = () => {
-  const currentConfig = {
-    debug: {
-      multisigKeys: {
-        '0xValidKey1': DevSecurityLevel.High,
-        '0xValidKey2': DevSecurityLevel.Medium
-      }
-    }
+const setup = (
+  multisigKeys?: any,
+  multiSigPermissions?: any
+) => {
+  multisigKeys = multisigKeys || {
+    '0xValidKey1': DevSecurityLevel.High,
+    '0xValidKey2': DevSecurityLevel.Medium
   }
 
-  const multiSigPermissions = {
+  multiSigPermissions = multiSigPermissions || {
     changeDevKeyList: ['0xValidKey1', '0xInvalidKey1'],
     changeMultiSigKeyList: ['0xValidKey2', '0xInvalidKey2']
+  }
+
+  const currentConfig = {
+    debug: {
+      multisigKeys
+    }
   }
 
   // Clean the permissions
@@ -95,4 +100,138 @@ describe('multisig integration tests', () => {
     expect(result.isKeyChange).toBe(true);
     expect(result.permittedKeys.length).toBeGreaterThan(0);
   });
-}); 
+
+  it('cleanMultiSigPermissions should filter keys before isNonKeyChange uses them', () => {
+    const { currentConfig, multiSigPermissions, cleanedPermissions } = setup(undefined, {
+      changeNonKeyConfigs: ['0xValidKey1', '0xInvalidKey1', '0xValidKey2']
+    })
+  
+    // Verify that invalid keys were removed
+    expect(cleanedPermissions.changeNonKeyConfigs).toEqual(['0xValidKey1', '0xValidKey2']);
+
+    // Mock a transaction that changes non-key configs
+    const tx = {
+      config: JSON.stringify({
+        debug: {
+          someConfig: 'new-value' // Changed non-key config
+        }
+      }),
+      type: 'ChangeConfig',
+      from: '0xSender',
+      cycle: mockCycle,
+      timestamp: Date.now()
+    } as unknown as ChangeConfig;
+
+    // Call isNonKeyChange with cleaned permissions
+    const result = isNonKeyChange(tx, currentConfig, cleanedPermissions);
+
+    console.log('multiSigPermissions', multiSigPermissions)
+    console.log('tx', tx)
+    console.log('cleanedPermissions', cleanedPermissions)
+    console.log('currentConfig', currentConfig)
+    // Verify that only valid keys are in the permitted keys
+    console.log('result', result)
+    expect(result.isNonKeyChange).toBe(true);
+    expect(result.permittedKeys).toEqual(['0xValidKey1', '0xValidKey2']);
+  });
+
+  it('isNonKeyChange should work with empty permitted keys after cleaning', () => {
+    const { currentConfig, cleanedPermissions } = setup(undefined, {
+      changeNonKeyConfigs: ['0xInvalidKey1', '0xInvalidKey2']
+    })
+
+    // Verify that all keys were removed
+    expect(cleanedPermissions.changeNonKeyConfigs).toEqual([]);
+
+    // Mock a transaction that changes non-key configs
+    const tx = {
+      config: JSON.stringify({
+        debug: {
+          someConfig: 'new-value' // Changed non-key config
+        }
+      }),
+      type: 'ChangeConfig',
+      from: '0xSender',
+      cycle: mockCycle,
+      timestamp: Date.now()
+    } as unknown as ChangeConfig;
+
+    // Call isNonKeyChange with cleaned permissions
+    const result = isNonKeyChange(tx, currentConfig, cleanedPermissions);
+
+    // Verify that isNonKeyChange is true but permittedKeys is empty
+    expect(result.isNonKeyChange).toBe(true);
+    expect(result.permittedKeys).toEqual([]);
+  });
+
+  it('should return noChange when configs are identical', () => {
+    // Mock config with multisig keys
+    const currentConfig = {
+      debug: {
+        multisigKeys: {
+          '0xValidKey1': 2,
+          '0xValidKey2': 1
+        },
+        someConfig: 'value'
+      }
+    };
+
+    // Mock permissions with valid keys
+    const multiSigPermissions = {
+      changeNonKeyConfigs: ['0xValidKey1', '0xValidKey2']
+    };
+
+    // Mock a transaction with identical config
+    const tx = {
+      config: JSON.stringify(currentConfig), // Same as current config
+      type: 'ChangeConfig',
+      from: '0xSender',
+      cycle: mockCycle,
+      timestamp: Date.now()
+    } as unknown as ChangeConfig;
+
+    // Call isNonKeyChange
+    const result = isNonKeyChange(tx, currentConfig, multiSigPermissions);
+
+    // Verify that isNonKeyChange is false and permittedKeys is empty
+    expect(result.isNonKeyChange).toBe(false);
+    expect(result.permittedKeys).toEqual([]);
+  });
+
+  it('should handle simultaneous key and non-key changes correctly', () => {
+    const currentConfig = {
+      debug: {
+        multisigKeys: {
+          '0xValidKey1': 2,
+          '0xValidKey2': 1
+        },
+        someConfig: 'value'
+      }
+    };
+
+    const tx = {
+      config: JSON.stringify({
+        debug: {
+          multisigKeys: {
+            '0xValidKey1': 2,
+            '0xValidKey3': 1  // Key change
+          },
+          someConfig: 'new-value'  // Non-key change
+        }
+      }),
+      type: 'ChangeConfig',
+      from: '0xSender',
+      cycle: mockCycle,
+      timestamp: Date.now()
+    } as unknown as ChangeConfig;
+
+    const multiSigPermissions = {
+      changeNonKeyConfigs: ['0xValidKey1', '0xValidKey2']
+    };
+
+    const result = isNonKeyChange(tx, currentConfig, multiSigPermissions);
+    // Should return false because key changes take precedence
+    expect(result.isNonKeyChange).toBe(false);
+    expect(result.permittedKeys).toEqual([]);
+  });
+});
